requirements.txt
Flask>=2.0
gspread>=5.0.0
google-auth>=2.0.0
# For local development with .env files (optional, but good practice)
# python-dotenv 
